[PRDP-17] Changed "Client" classes to implementations of interfaces
